wip/home meeting (#128)
* basic

* model

* 更新

* 更新

* edit

* 更改createSheep

* wip

* 更新

* model

* 更新

* 更新

* wip

* model

* homeMeetingRecordId

* 更新

* 更新

* e/ vup

* wip

+++ controllers/UserSerialNumber/DeleteSerialNumberController.ts
@@ -3,7 +3,15 @@ import { UserSerialNumber } from '../../models/UserSerialNumber'
 import { Req, Res } from '../../types'
 
 export const DeleteSerialNumberController = async (req: Req, res: Res) => {
-  const { id } = req.params
+  const { id } = req.body
+
+  if (!id) {
+    res.status(StatusCodes.BAD_REQUEST).json({
+      msg: 'User serial number id is required'
+    })
+    return
+  }
+  
   await UserSerialNumber.findByIdAndDelete(id)
   res.status(StatusCodes.OK).json({
     msg: 'User serial number deleted successfully'
